Added authenticate_probable_user.
If someone requests a /shares/ or /shares/x they probably are already a
user. (They may be following a share notification link from their
email.)

If they do not pass authentication because they do not have an active
session, they should be taken to the log-in page instead of the
homepage.