After logging in, redirect to original destination.
If a user attempts to access a page which requires authentication,
then their destination is stored. When they log in successfully,
they are sent to their original destination.

Based (heavily) on Authlogic's example:
  http://github.com/binarylogic/authlogic_example

Also, the authenticate_probable_user from my previous commit is gone.
All failed authentications now go to the log-in page. I figure anyone
with an "internal" URL is probably a user. And just in case they aren't,
I added a link to create an account on the log-in page.